+++ packages/std/src/results/mod.rs
@@ -14,9 +14,9 @@ pub use attribute::{attr, Attribute};
 #[allow(deprecated)]
 pub use context::Context;
 pub use contract_result::ContractResult;
-#[cfg(feature = "staking")]
-pub use cosmos_msg::StakingMsg;
 pub use cosmos_msg::{wasm_execute, wasm_instantiate, BankMsg, CosmosMsg, WasmMsg};
+#[cfg(feature = "staking")]
+pub use cosmos_msg::{DistributionMsg, StakingMsg};
 pub use empty::Empty;
 pub use query::QueryResponse;
 pub use response::Response;